Refactor cache to own package
