+++ internal/cache/cache.go
@@ -1,4 +1,4 @@
-package aurora_test_auth
+package cache
 
 import (
 	"context"
@@ -11,4 +11,3 @@ type Cache interface {
 	Get(ctx context.Context, key string) (string, error)
 	Close(ctx context.Context) error
 }
-
